requirements.txt
openai==0.27.8
langchain==0.0.352
python-dotenv
chromaDB